refactor: make e2e tests more maintainable
