+++ packages/diracx-web/cypress.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "cypress";
 export default defineConfig({
   e2e: {
     specPattern: "test/e2e/**/*.cy.ts",
-    supportFile: false,
+    supportFile: "test/e2e/support/e2e.ts",
     setupNodeEvents(_on, _config) {
       // implement node event listeners here
     },
