+++ packages/diracx-web/test/e2e/importExportState.cy.ts
@@ -1,20 +1,9 @@
 /// <reference types="cypress" />
+/// <reference path="support/index.d.ts" />
 
 describe("Export and import app state", () => {
   beforeEach(() => {
-    cy.session("login", () => {
-      cy.visit("/");
-      //login
-      cy.get('[data-testid="login-form-button"]').click();
-      cy.get("#login").type("admin@example.com");
-      cy.get("#password").type("password");
-
-      // Find the login button and click on it
-      cy.get("button").click();
-      // Grant access
-      cy.get(":nth-child(1) > form > .dex-btn").click();
-      cy.url().should("include", "/auth");
-    });
+    cy.login();
 
     cy.visit("/");
