+++ packages/diracx-web/test/e2e/support/e2e.ts
@@ -1,0 +1,1 @@
+import "./commands";
