+++ packages/diracx-web/test/e2e/support/index.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="cypress" />
+
+declare namespace Cypress {
+  interface Chainable {
+    /**
+     * Log in via the OIDC provider and cache the session.
+     * Uses cy.session() so the login flow only runs once per spec.
+     */
+    login(): Chainable<void>;
+  }
+}
